split game states
