+++ src/overlays/gamestates/ovl_select/z_select.c
@@ -4,17 +4,28 @@
  * Description: Debug Scene Select Menu
  */
 
+#include "libc64/qrand.h"
+#include "libu64/gfxprint.h"
 #include "ultra64.h"
+#include "controller.h"
+#include "gfx.h"
+#include "gfx_setupdl.h"
 #include "alloca.h"
 #if PLATFORM_N64
 #include "n64dd.h"
 #endif
+#include "regs.h"
+#include "segment_symbols.h"
 #include "seqcmd.h"
 #include "sequence.h"
+#include "sfx.h"
 #include "terminal.h"
+#include "z_select.h"
+#include "z_title.h"
+#include "z64play.h"
 #include "z64save.h"
 
-#include "global.h"
+#include "macros.h"
 
 void MapSelect_LoadTitle(MapSelectState* this) {
     this->state.running = false;
